Remove emitting id when writing FHIR resource. Update invocation toolkit to load resources.
