+++ Cql/Cql.Firely/PublicAPI.Unshipped.txt
@@ -1,4 +1,5 @@
 ﻿#nullable enable
+Hl7.Cql.Fhir.Extensions.FhirDeserializationExtensions
 Hl7.Cql.Fhir.FhirCqlContext
 Hl7.Cql.Fhir.FhirCqlContextOptions
 Hl7.Cql.Fhir.FhirCqlContextOptions.CodeInOperatorSemantics
@@ -21,6 +22,9 @@ Hl7.Cql.Fhir.FhirTypeResolver
 Hl7.Cql.Fhir.FhirTypeResolver.FhirTypeResolver(Hl7.Fhir.Introspection.ModelInspector! inspector) -> void
 Hl7.Cql.Fhir.ValueSetExtensions
 override Hl7.Cql.Fhir.FhirTypeResolver.GetPropertyCore(System.Type! type, string! propertyName) -> System.Reflection.PropertyInfo?
+static Hl7.Cql.Fhir.Extensions.FhirDeserializationExtensions.DeserializeJsonToFhir<TDomainResource>(this string! jsonString) -> TDomainResource!
+static Hl7.Cql.Fhir.Extensions.FhirDeserializationExtensions.DeserializeJsonToFhir<TDomainResource>(this System.IO.Stream! jsonStream) -> TDomainResource!
+static Hl7.Cql.Fhir.Extensions.FhirDeserializationExtensions.DeserializeJsonToFhir<TDomainResource>(this System.ReadOnlySpan<char> jsonCharSpan) -> TDomainResource!
 static Hl7.Cql.Fhir.FhirCqlContext.ForBundle(Hl7.Fhir.Model.Bundle? bundle = null, System.Collections.Generic.IDictionary<string!, object!>? parameters = null, Hl7.Cql.ValueSets.IValueSetDictionary? valueSets = null, System.DateTimeOffset? now = null, Hl7.Cql.Runtime.DefinitionDictionary<System.Delegate!>? delegates = null, Hl7.Cql.Fhir.FhirCqlContextOptions? options = null) -> Hl7.Cql.Runtime.CqlContext!
 static Hl7.Cql.Fhir.FhirCqlContext.WithDataSource(Hl7.Cql.Operators.IDataSource? source = null, System.Collections.Generic.IDictionary<string!, object!>? parameters = null, Hl7.Cql.ValueSets.IValueSetDictionary? valueSets = null, System.DateTimeOffset? now = null, Hl7.Cql.Runtime.DefinitionDictionary<System.Delegate!>? delegates = null, Hl7.Cql.Fhir.FhirCqlContextOptions? options = null) -> Hl7.Cql.Runtime.CqlContext!
 static Hl7.Cql.Fhir.FhirDeserializationExtensions.DeserializeJsonToFhir<TDomainResource>(this string! jsonString) -> TDomainResource!
