+++ Cql/Cql.Invocation/PublicAPI.Unshipped.txt
@@ -42,6 +42,7 @@ static Hl7.Cql.Invocation.Toolkit.Extensions.ElmToolkitExtensions.CreateLibraryS
 static Hl7.Cql.Invocation.Toolkit.Extensions.InvocationToolkitExtensions.AddAssemblyBinaries(this Hl7.Cql.Invocation.Toolkit.InvocationToolkit! invocationToolkit, params Hl7.Cql.CodeGeneration.NET.AssemblyBinary![]! assemblyBinary) -> Hl7.Cql.Invocation.Toolkit.InvocationToolkit!
 static Hl7.Cql.Invocation.Toolkit.Extensions.InvocationToolkitExtensions.AddAssemblyBinaryFiles(this Hl7.Cql.Invocation.Toolkit.InvocationToolkit! invocationToolkit, System.Collections.Generic.IEnumerable<System.IO.FileInfo!>! files) -> Hl7.Cql.Invocation.Toolkit.InvocationToolkit!
 static Hl7.Cql.Invocation.Toolkit.Extensions.InvocationToolkitExtensions.AddAssemblyBinaryFilesFromDirectory(this Hl7.Cql.Invocation.Toolkit.InvocationToolkit! invocationToolkit, System.IO.DirectoryInfo! directory, System.IO.EnumerationOptions? options = null, System.Func<System.IO.FileInfo!, bool>? filePredicate = null) -> Hl7.Cql.Invocation.Toolkit.InvocationToolkit!
+static Hl7.Cql.Invocation.Toolkit.Extensions.InvocationToolkitExtensions.AddFhirLibrariesFromDirectory(this Hl7.Cql.Invocation.Toolkit.InvocationToolkit! invocationToolkit, System.IO.DirectoryInfo! directory, System.IO.EnumerationOptions? options = null, System.Func<System.IO.FileInfo!, bool>? filePredicate = null) -> Hl7.Cql.Invocation.Toolkit.InvocationToolkit!
 static Hl7.Cql.Invocation.Toolkit.Extensions.LibraryInvokerExtensions.EnumerateLibraryDefinitionsResults(this Hl7.Cql.Invocation.Toolkit.LibraryInvoker! libraryInvoker, Hl7.Cql.Runtime.CqlContext! cqlContext, System.Func<Hl7.Cql.Invocation.Toolkit.DefinitionInvoker!, bool>? includeDefinition = null, Hl7.Cql.Runtime.ValueExceptionHandler<Hl7.Cql.Invocation.Toolkit.DefinitionInvoker!>? definitionInvokerExceptionHandler = null) -> System.Collections.Generic.IEnumerable<(Hl7.Cql.Invocation.Toolkit.DefinitionInvoker! definitionInvoker, object? definitionResult)>!
 static Hl7.Cql.Invocation.Toolkit.Extensions.LibrarySetInvokerExtensions.EnumerateLibraryDefinitionsResults(this Hl7.Cql.Invocation.Toolkit.LibrarySetInvoker! librarySetInvoker, Hl7.Cql.Runtime.CqlContext! cqlContext, Hl7.Cql.Runtime.CqlVersionedLibraryIdentifier libraryIdentifier, System.Func<Hl7.Cql.Invocation.Toolkit.DefinitionInvoker!, bool>? includeDefinition = null, Hl7.Cql.Runtime.ValueExceptionHandler<Hl7.Cql.Invocation.Toolkit.DefinitionInvoker!>? definitionInvokerExceptionHandler = null) -> System.Collections.Generic.IEnumerable<(Hl7.Cql.Invocation.Toolkit.DefinitionInvoker! definitionInvoker, object? definitionResult)>!
 static Hl7.Cql.Invocation.Toolkit.Extensions.LibrarySetInvokerExtensions.EnumerateLibrarySetDefinitionsResults(this Hl7.Cql.Invocation.Toolkit.LibrarySetInvoker! librarySetInvoker, Hl7.Cql.Runtime.CqlContext! cqlContext, System.Func<Hl7.Cql.Invocation.Toolkit.DefinitionInvoker!, bool>? includeDefinition = null, Hl7.Cql.Runtime.ValueExceptionHandler<Hl7.Cql.Invocation.Toolkit.DefinitionInvoker!>? definitionInvokerExceptionHandler = null) -> System.Collections.Generic.IEnumerable<(Hl7.Cql.Invocation.Toolkit.DefinitionInvoker! definitionInvoker, object? definitionResult)>!
