+++ Cql/Cql.Packaging/DataRequirementsAnalyzer.cs
@@ -22,7 +22,7 @@ namespace Hl7.Cql.Packaging;
 /// Resources used by the retrieves. There's much more to it, as can be glanced
 /// from the public Java version here: https://github.com/cqframework/clinical_quality_language/blob/master/Src/java/elm-fhir/src/main/java/org/cqframework/cql/elm/requirements/fhir/DataRequirementsProcessor.java
 /// </remarks>
-internal class DataRequirementsAnalyzer(ElmLibrarySet librarySet, Elm.Library focusLibrary)
+internal class DataRequirementsAnalyzer(ElmLibrarySet librarySet, ElmLibrary focusLibrary)
 {
     /// <summary>
     /// Visits the ELM in the LibrarySet and extracts the DataRequirements from it.
@@ -35,7 +35,7 @@ internal class DataRequirementsAnalyzer(ElmLibrarySet librarySet, Elm.Library fo
         return result;
     }
 
-    private void Visit(Elm.Library library, List<DataRequirement> allRequirements)
+    private void Visit(ElmLibrary library, List<DataRequirement> allRequirements)
     {
         var walker = new Elm.ElmTreeWalker(n => Visit(library, allRequirements, n));
         walker.Start(library);
@@ -48,7 +48,7 @@ internal class DataRequirementsAnalyzer(ElmLibrarySet librarySet, Elm.Library fo
     }
 
     private bool Visit(
-        Elm.Library library,
+        ElmLibrary library,
         List<DataRequirement> result,
         object node)
     {
@@ -112,7 +112,7 @@ internal class DataRequirementsAnalyzer(ElmLibrarySet librarySet, Elm.Library fo
     //   private static string ToReference(Elm.CodeSystemDef def) => def.id + (def.version is { } v ? $"|{v}" : null);
 
 
-    private class CodeFilterComponentBuilder(ElmLibrarySet librarySet, Elm.Library contextLibrary)
+    private class CodeFilterComponentBuilder(ElmLibrarySet librarySet, ElmLibrary contextLibrary)
     {
         public DataRequirement.CodeFilterComponent ToCodeFilterComponent(
             string property,
