+++ Cql/Cql.Runtime/Cql.Runtime.csproj
@@ -28,6 +28,7 @@
 		<InternalsVisibleTo Include="CqlToElmTests" Key="$(LibraryPKHash)" />
 		<InternalsVisibleTo Include="Hl7.Cql.Elm" Key="$(LibraryPKHash)" />
 		<InternalsVisibleTo Include="Hl7.Cql.MultiPackager" Key="$(LibraryPKHash)" />
+		<!-- REVIEW: The line below must be removed before a merge -->
 		<InternalsVisibleTo Include="IntegrationRunner" Key="$(LibraryPKHash)" />
 	</ItemGroup>
