+++ Demo/Test.Measures.Demo/MeasuresTest.cs
@@ -9,6 +9,7 @@ using Hl7.Cql.Runtime;
 using Hl7.Cql.CodeGeneration.NET.Toolkit;
 using Hl7.Cql.Invocation.Toolkit;
 using Hl7.Cql.Invocation.Toolkit.Extensions;
+using Hl7.Cql.Fhir.Extensions;
 
 namespace Test
 {
